+++ packages/smarthr-ui/src/components/Select/client/index.ts
@@ -1,0 +1,2 @@
+export { ActualSelect, NotOmittingLabelsInMobileSafari } from './ActualSelect'
+export type { Props as ActualSelectProps } from './ActualSelect'
